requirements.txt
#telethon
#pyyaml
numpy
pandas
selenium
sqlalchemy==1.3.23
pymysql==1.0.2
#mysqlclient==2.0.3
mysql-connector-python==8.0.22
telethon